PutFollow (delete) is now handled through AWS and fully async. (still a mess)

+++ app/src/main/java/edu/byu/cs/tweeter/view/ui/main/storyView/StoryViewActivity.java
@@ -28,12 +28,13 @@ import edu.byu.cs.tweeter.view.asyncTasks.DeleteFollowTask;
 import edu.byu.cs.tweeter.view.asyncTasks.FindUserTask;
 import edu.byu.cs.tweeter.view.asyncTasks.GetFollowTask;
 import edu.byu.cs.tweeter.view.asyncTasks.LoadImageTask;
+import edu.byu.cs.tweeter.view.asyncTasks.PutFollowTask;
 import edu.byu.cs.tweeter.view.cache.ImageCache;
 
 /**
  * The main activity for the application. Contains tabs for story, following, and followers.
  */
-public class StoryViewActivity extends AppCompatActivity implements LoadImageTask.LoadImageObserver, StoryViewPresenter.View, GetFollowTask.GetFollowObserver, FindUserTask.FindUserObserver,FindUserPresenter.View, DeleteFollowTask.GetFollowObserver {
+public class StoryViewActivity extends AppCompatActivity implements LoadImageTask.LoadImageObserver, StoryViewPresenter.View, GetFollowTask.GetFollowObserver, FindUserTask.FindUserObserver,FindUserPresenter.View, DeleteFollowTask.GetFollowObserver, PutFollowTask.GetFollowObserver {
     private static final String LOG_TAG = "StoryViewActivity";
 
     private StoryViewActivity storyViewActivity;
@@ -83,28 +84,16 @@ public class StoryViewActivity extends AppCompatActivity implements LoadImageTas
 
         followButton.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
-                DeleteFollowTask deleteFollowTask = new DeleteFollowTask(presenter, storyViewActivity);
-
                 FollowRequest request = new FollowRequest(presenter.getCurrentUser().getAlias(), user.getAlias());
-                deleteFollowTask.execute(request);
 
-//
-//                ServerFacade serverFacade = new ServerFacade();
-//                if(isFollowing){
-//                    //Remove following relation
-//
-//
-////todo clean up
-////                    serverFacade.removeFollowing(user, presenter.getCurrentUser()); //todo make this async
-////                    followButton.setText(R.string.Follow);
-////                    isFollowing = false;
-//                }
-//                else{
-//                    //Add following relation
-//                    serverFacade.addFollowing(user, presenter.getCurrentUser()); //todo make this async
-//                    followButton.setText(R.string.Following);
-//                    isFollowing = true;
-//                }
+                if(isFollowing){ //Remove following relation
+                    DeleteFollowTask deleteFollowTask = new DeleteFollowTask(presenter, storyViewActivity);
+                    deleteFollowTask.execute(request);
+                }
+                else{ //Add following relation
+                    PutFollowTask putFollowTask = new PutFollowTask(presenter, storyViewActivity);
+                    putFollowTask.execute(request);
+                }
             }
         });
 
@@ -158,11 +147,11 @@ public class StoryViewActivity extends AppCompatActivity implements LoadImageTas
     @Override
     public void followRetrieved(FollowResponse followResponse) {
         if(followResponse.isFollows()){
-            followButton.setText("Following");
+            followButton.setText(R.string.Following);
             isFollowing = true;
         }
         else{
-            followButton.setText("Follow");
+            followButton.setText(R.string.Follow);
             isFollowing = false;
         }
     }
